+++ app/src/main/java/org/ole/planet/myplanet/ui/userprofile/BaseAchievementFragment.java
@@ -1,18 +1,32 @@
 package org.ole.planet.myplanet.ui.userprofile;
 
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.text.TextUtils;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.CheckBox;
+import android.widget.CheckedTextView;
 import android.widget.EditText;
 import android.widget.LinearLayout;
+import android.widget.ListView;
 import android.widget.TextView;
 
 import com.google.gson.JsonArray;
 
+import org.ole.planet.myplanet.R;
 import org.ole.planet.myplanet.base.BaseContainerFragment;
 import org.ole.planet.myplanet.model.RealmAchievement;
+import org.ole.planet.myplanet.model.RealmMyLibrary;
 import org.ole.planet.myplanet.model.RealmUserModel;
+import org.ole.planet.myplanet.utilities.CheckboxListView;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import io.realm.Realm;
 
@@ -47,6 +61,34 @@ public class BaseAchievementFragment extends BaseContainerFragment {
         etLastName.setText(user.getLastName());
         etBirthPlace.setText(user.getBirthPlace());
     }
+    public CheckboxListView createResourceList(View v, List<RealmMyLibrary> list, List<String> prevList) {
+
+        CheckboxListView lv = v.findViewById(R.id.alertDialog_listView);
+        ArrayList<String> names = new ArrayList<>();
+        ArrayList<Integer> selected = new ArrayList();
+        for (int i = 0; i < list.size(); i++) {
+            names.add(list.get(i).getTitle());
+            if (prevList.contains(list.get(i).getTitle()))
+                selected.add(i);
+        }
+
+        ArrayAdapter<String> adapter = new ArrayAdapter<String>(getActivity(), R.layout.item_checkbox, R.id.checkBoxRowLayout, names) {
+            @NonNull
+            @Override
+            public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
+                CheckedTextView textView = (CheckedTextView) LayoutInflater.from(getActivity()).inflate(R.layout.rowlayout, parent, false);
+                textView.setText(getItem(position));
+                textView.setChecked(lv.getSelectedItemsList().contains(position));
+                lv.setItemChecked(position, lv.getSelectedItemsList().contains(position));
+                return textView;
+            }
+        };
+        lv.setSelectedItemsList(selected);
+        lv.setChoiceMode(ListView.CHOICE_MODE_MULTIPLE);
+        lv.setAdapter(adapter);
+        return lv;
+    }
+
 
 
     public void setUserInfo() {
